requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.72
langchain-openai==0.3.28
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.11
openai==1.99.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
zstandard==0.23.0
